requirements.txt
appdirs==1.4.4
beautifulsoup4==4.12.2
click==8.1.7
contourpy==1.1.1
cycler==0.12.1
docker-pycreds==0.4.0
einops==0.7.0
faiss-gpu==1.7.2
filelock==3.12.4
fonttools==4.43.1
gdown==4.7.1
gitdb==4.0.11
gitpython==3.1.40
importlib-resources==6.1.0
infomap==2.7.1
jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
loguru==0.7.2
markupsafe==2.1.3
matplotlib==3.8.0
opencv-python==4.8.1.78
packaging==23.2
pandas==2.1.1
path==16.7.1
pathtools==0.1.2
protobuf==4.24.4
psutil==5.9.6
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
pyyaml==6.0.1
scikit-learn==1.3.2
scipy==1.11.3
sentry-sdk==1.32.0
setproctitle==1.3.3
simplejson==3.19.2
six==1.16.0
smmap==5.0.1
soupsieve==2.5
threadpoolctl==3.2.0
torch-geometric==2.4.0
tqdm==4.66.1
tzdata==2023.3
wandb==0.15.12
warmup-scheduler==0.3
zipp==3.17.0